+++ petl/errors.py
@@ -17,3 +17,12 @@ class FieldSelectionError(Exception):
 
     def __str__(self):
         return 'selection is not a field or valid field index: %r' % self.value
+
+
+class ArgumentError(Exception):
+
+    def __init__(self, message):
+        self.message = message
+
+    def __str__(self):
+        return 'argument error: %s' % self.message