+++ petl/test/transform/test_regex.py
@@ -174,6 +174,25 @@ def test_search():
     ieq(expect3, table3)
 
 
+def test_search_2():
+
+    # test ported from selectre
+    table = (('foo', 'bar', 'baz'),
+             ('aa', 4, 9.3),
+             ('aaa', 2, 88.2),
+             ('b', 1, 23.3),
+             ('ccc', 8, 42.0),
+             ('bb', 7, 100.9),
+             ('c', 2))
+    actual = search(table, 'foo', '[ab]{2}')
+    expect = (('foo', 'bar', 'baz'),
+              ('aa', 4, 9.3),
+              ('aaa', 2, 88.2),
+              ('bb', 7, 100.9))
+    ieq(expect, actual)
+    ieq(expect, actual)
+
+
 def test_searchcomplement():
 
     table1 = (('foo', 'bar', 'baz'),